requirements.txt
django>=5.0,<6.0
gunicorn
whitenoise[broccoli]
dj-database-url
PyMySQL
python-dotenv
